# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Vakil & Co. Legal Services
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    post:
      operationId: createContact
      tags: [contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listContacts
      tags: [contacts]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/services-config:
    get:
      operationId: listServicesConfig
      tags: [admin-services]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceConfig"
    post:
      operationId: createServiceConfig
      tags: [admin-services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceConfigBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceConfig"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/services-config/{id}:
    patch:
      operationId: updateServiceConfig
      tags: [admin-services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceConfigBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceConfig"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteServiceConfig
      tags: [admin-services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin: Quotations ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Consultation ──────────────────────────────────────────────────────────
...
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
    UpdateContactBody:
      type: object
      properties:
        status:
          type: string
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - subject
        - message
        - status
        - createdAt
        - updatedAt

    # ── Newsletter ────────────────────────────────────────────────────────────
...
    ServiceConfig:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: string
        serviceName:
          type: string
        displayName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        basePrice:
          type: ["integer", "null"]
        priceDisplay:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        isPopular:
          type: boolean
        isActive:
          type: boolean
        updatedAt:
          type: string
      required:
        - id
        - categoryId
        - serviceName
        - isPopular
        - isActive
        - updatedAt
    CreateServiceConfigBody:
      type: object
      properties:
        categoryId:
          type: string
        serviceName:
          type: string
        displayName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        basePrice:
          type: ["integer", "null"]
        priceDisplay:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        isPopular:
          type: boolean
        isActive:
          type: boolean
      required:
        - categoryId
        - serviceName
    UpdateServiceConfigBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        basePrice:
          type: ["integer", "null"]
        priceDisplay:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        isPopular:
          type: boolean
        isActive:
          type: boolean

    # ── Quotations ────────────────────────────────────────────────────────────
...