# lib/api-zod/src/generated/types/updateContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Vakil & Co. Legal Services
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateContactBody {
  status?: string;
}